v.boxplot and r.boxplot: plot/layout options added / changed (#1271)
* v.boxplot and r.boxplot: plot/layout options added / changed

Following suggestions in #1071
- Adding plot and layout options to v.boxplot
- Changes in parameter names of v.boxplot and r.boxplot

* v.boxplot: add layout/plot options

- Add option
  - to set limits of value axis
  - add grid lines
- Reorganize tabs
- Lazy loading matplotlib
- Fix: selected colors checked and changed to matplotlib format

* v.boxplot and r.boxplot: solve issue with running addons in VS Code

Running the addons in VS Code results in an error related with `matplotlib.use("WXAgg")`. A possible solution is using another backend like `WebAgg`. However, running VS Code probably means one wants to print the resulting graph to file. As in that case, the backend is not needed, we can just avoid loading it when the option to save the graph to file is selected.

* correction code